+++ backend/registry/mapping/capabilities/wms/wms130.py
@@ -1,13 +1,11 @@
-from copy import deepcopy
-
 from odin.mapping import Mapping, assign_field, map_field
-from ows_lib.xml_mapper.capabilities.mixins import OperationUrl
 from ows_lib.xml_mapper.capabilities.wms.wms130 import Layer as XmlLayer
 from ows_lib.xml_mapper.capabilities.wms.wms130 import \
     RemoteMetadata as XmlRemoteMetadata
 from ows_lib.xml_mapper.capabilities.wms.wms130 import Style as XmlStyle
 from ows_lib.xml_mapper.capabilities.wms.wms130 import \
     WebMapService as XmlWebMapService
+from registry.mapping.capabilities.mixins import OgcServiceToXmlMixin
 from registry.models.metadata import DatasetMetadata, Style
 from registry.models.service import Layer, WebMapService
 
@@ -59,41 +57,21 @@ class LayerToXml(Mapping):
         return _remote_metadata
 
 
-class WebMapServiceToXml(Mapping):
+class WebMapServiceToXml(OgcServiceToXmlMixin, Mapping):
     from_obj = WebMapService
     to_obj = XmlWebMapService
 
-    def __init__(self, destination_obj: XmlWebMapService, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.xml = deepcopy(destination_obj)
-
     def update(self, *args, **kwargs):
-        updated_service = super().update(destination_obj=self.xml, *args, **kwargs)
+        updated_service = super().update(*args, **kwargs)
         self._update_layers()
-        self._update_operation_urls(updated_service=updated_service)
         return updated_service
 
     def _update_layers(self):
         for layer in self.source.layers.filter(is_active=True):
-            LayerToXml(source_obj=layer, destination_obj=self.xml.get_layer_by_identifier(
+            LayerToXml(source_obj=layer, destination_obj=self.destination_obj.get_layer_by_identifier(
                 identifier=layer.identifier)).update()
         for layer in self.source.layers.filter(is_active=False):
-            xml_layer = self.xml.get_layer_by_identifier(
+            xml_layer = self.destination_obj.get_layer_by_identifier(
                 identifier=layer.identifier)
             if xml_layer:
                 del xml_layer
-
-    def _update_operation_urls(self, updated_service):
-        operation_urls = []
-        for operation_url in self.source.operation_urls.all():
-            operation_urls.append(
-                OperationUrl(
-                    method=operation_url.method,
-                    url=operation_url.url,
-                    operation=operation_url.operation,
-                    mime_types=[str(mime_type) for mime_type in operation_url.mime_types.all()]))
-        updated_service.operation_urls = operation_urls
-
-    @assign_field(to_list=True)
-    def keywords(self):
-        return [str(keyword) for keyword in self.source.keywords.all()]
